+++ packages/clerk-js/src/ui/components/UserProfile/__tests__/MfaPage.test.tsx
@@ -244,6 +244,7 @@ describe('MfaPage', () => {
 
       expect(itemButton).toBeDefined();
       await act(async () => {
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         await userEvent.click(itemButton!);
       });
       await waitFor(() => getByText(/^regenerate$/i));
@@ -294,6 +295,7 @@ describe('MfaPage', () => {
       expect(itemButton).toBeDefined();
 
       await act(async () => {
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         await userEvent.click(itemButton!);
       });
       await waitFor(() => getByText(/^remove$/i));
@@ -327,6 +329,7 @@ describe('MfaPage', () => {
       expect(itemButton).toBeDefined();
 
       await act(async () => {
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         await userEvent.click(itemButton!);
       });
       await waitFor(() => getByText(/^remove$/i));
@@ -371,6 +374,7 @@ describe('MfaPage', () => {
       expect(itemButton).toBeDefined();
 
       await act(async () => {
+        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         await userEvent.click(itemButton!);
       });
       await waitFor(() => getByText(/^remove$/i));